1、用户模块增删异常bug修复。 2、状态模块偏差保留两位小数显示。

+++ public/javascripts/controllers/users.js
@@ -1,54 +1,43 @@
 /**
  * Created by Administrator on 2017/7/23.
  */
-define(['services/usersService'], function(){
-	return['$scope', 'UsersService', function($scope, usersService){
-		var oldUsers = [];
-		usersService.getUsers().then(function(rs){
+define(['services/usersService'], function () {
+	return ['$scope', 'UsersService', function ($scope, usersService) {
+		usersService.getUsers().then(function (rs) {
 			$scope.users = rs.data;
 			console.log(rs.data);
-			oldUsers = $scope.users;
 		});
-		$scope.addUser = function(){
-			$scope.AddUser = {
+		$scope.addUser = function () {
+			$scope.newUser = {
 				Username: '',
 				Password: '',
 				IsActive: true,
 				IsAPI: false,
 				IsNew: true
 			};
-			$scope.users.push($scope.AddUser);
+			$scope.users.push($scope.newUser);
 		};
-		$scope.removeUser = function(index){
-			usersService.deleteUser({params:{id: $scope.users[index].Username}}).then(function(rs){
-				console.log(rs);
-			});
-			usersService.getUsers().then(function(rs){
-				$scope.users = rs.data;
-				oldUsers = rs.data;
-			});
-		};
-		$scope.saveOrUpdateUser = function(user, index){
-			if(user.IsNew){
-				delete user.IsNew;
-				usersService.saveUser(user).then(function(rs){
+		$scope.removeUser = function (index) {
+			if (!$scope.users[index].IsNew) {
+				usersService.deleteUser({params: {id: $scope.users[index].Username}}).then(function (rs) {
 					console.log(rs);
 				});
-				usersService.getUsers().then(function(rs){
-					$scope.users = rs.data;
-					oldUsers = rs.data;
-				});
 			}
-			else{
+			$scope.users.splice(index, 1);
+		};
+		$scope.saveOrUpdateUser = function (user, index) {
+			if (user.IsNew) {
 				delete user.IsNew;
-				usersService.updateUser(user, {params: {id: oldUsers[index].Username}}).then(function(rs){
+				usersService.saveUser(user).then(function (rs) {
 					console.log(rs);
 				});
-				usersService.getUsers().then(function(rs){
-					$scope.users = rs.data;
-					oldUsers = rs.data;
+			} else {
+				delete user.IsNew;
+				usersService.updateUser(user, user.Username).then(function (rs) {
+					console.log(rs);
 				});
 			}
+			$scope.users[index] = user;
 		};
 	}];
 })
